frontend/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
mysql-connector-python==8.2.0
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0
fastapi-websocket==0.1.7
python-socketio==5.10.0
cors==1.0.1
fastapi-cors==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2